backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
pydantic==2.7.1
pydantic-settings==2.2.1
python-dotenv==1.0.1
httpx==0.27.0
openai==1.30.1
langchain==0.2.1
langchain-openai==0.1.8
langchain-community==0.2.1
networkx==3.3
scipy==1.13.0
pytest==8.2.0
pytest-asyncio==0.23.7
rich==13.7.1
# Database
sqlalchemy==2.0.30
aiosqlite==0.20.0
# Auth
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
# RAG
chromadb==0.5.3
sentence-transformers==3.0.1
